Move `FindUserResult` to the `user.finduser` package
