+++ app/src/main/java/org/simple/clinic/user/UserSession.kt
@@ -17,9 +17,8 @@ import org.simple.clinic.facility.FacilityRepository
 import org.simple.clinic.forgotpin.ForgotPinResponse
 import org.simple.clinic.forgotpin.ResetPinRequest
 import org.simple.clinic.login.LoginApi
-import org.simple.clinic.login.LoginResponse
 import org.simple.clinic.patient.PatientRepository
-import org.simple.clinic.registration.FindUserResult
+import org.simple.clinic.user.finduser.FindUserResult
 import org.simple.clinic.registration.RegistrationApi
 import org.simple.clinic.registration.RegistrationRequest
 import org.simple.clinic.registration.RegistrationResponse
