+++ app/src/test/java/org/simple/clinic/registration/phone/RegistrationPhoneScreenControllerTest.kt
@@ -20,7 +20,7 @@ import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.simple.clinic.patient.PatientMocker
-import org.simple.clinic.registration.FindUserResult
+import org.simple.clinic.user.finduser.FindUserResult
 import org.simple.clinic.registration.phone.PhoneNumberValidator.Result.LENGTH_TOO_SHORT
 import org.simple.clinic.registration.phone.PhoneNumberValidator.Result.VALID
 import org.simple.clinic.registration.phone.PhoneNumberValidator.Type.MOBILE
